% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{defaultOpenCVPath}
\alias{defaultOpenCVPath}
\title{Default Install Location of OpenCV}
\usage{
defaultOpenCVPath()
}
\value{
A character string.
}
\description{
This function returns the location at which OpenCV should be
 installed by default.
}
\examples{
\dontrun{
 defaultOpenCVPath()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
