% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{isCmakeInstalled}
\alias{isCmakeInstalled}
\title{Check Cmake Installation}
\usage{
isCmakeInstalled()
}
\value{
A boolean indicating whether Cmake was or not installed on the system.
}
\description{
This functions checks that Cmake is installed on the system.
}
\examples{
isCmakeInstalled()

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
