% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{isOpenCVInstalled}
\alias{isOpenCVInstalled}
\title{Check OpenCV Installation}
\usage{
isOpenCVInstalled()
}
\value{
A boolean.
}
\description{
This function checks that OpenCV is installed and accessible.
}
\examples{
isOpenCVInstalled()

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
