% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{OpenCVPath}
\alias{OpenCVPath}
\title{Install Location of OpenCV}
\usage{
OpenCVPath()
}
\value{
A character string.
}
\description{
This function returns the location at which OpenCV is installed.
}
\examples{
\dontrun{
 OpenCVPath()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
