% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{removeOpenCV}
\alias{removeOpenCV}
\title{Remove OpenCV}
\usage{
removeOpenCV()
}
\value{
A boolean.
}
\description{
This function removes OpenCV from the system.
}
\examples{
\dontrun{
 installOpenCV()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
