% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installOpenCV}
\alias{installOpenCV}
\title{Install OpenCV}
\usage{
installOpenCV(
  install_path = defaultOpenCVPath(),
  batch = FALSE,
  use_ccache = FALSE,
  optimize_for_host = FALSE,
  modules = c("calib3d", "core", "dnn", "features2d", "flann", "gapi", "highgui",
    "imgcodecs", "imgproc", "ml", "objdetect", "photo", "stitching", "video", "videoio",
    "ximgproc", "wechat_qrcode")
)
}
\arguments{
\item{install_path}{A character string indicating the location at which
OpenCV should be installed. By default, it is the value returned by
\code{\link{defaultOpenCVPath}}.}

\item{batch}{A boolean indicating whether to skip (\code{TRUE}) or not
(\code{FALSE}, the default) the interactive installation dialog. This is
useful when OpenCV needs to be installed in a non-interactive environment
(e.g., during a batch installation on a server).}

\item{use_ccache}{A boolean indicating whether to use \code{ccache}
(\code{TRUE}) to speed up repeated compilations. Requires \code{ccache} to
be installed and on the PATH. Defaults to \code{FALSE}.}

\item{optimize_for_host}{A boolean indicating whether to let CMake detect
and use the host CPU's full instruction set (\code{TRUE}) instead of the
default portable build targeting SSE4/AVX (\code{FALSE}). Produces a faster
binary on the build machine but the result may not run on other hardware.}

\item{modules}{A character vector of OpenCV modules to compile. Defaults to
the full set supported by \code{ROpenCVLite}. Specify a subset to reduce
compilation time when only specific functionality is needed.}
}
\value{
A boolean.
}
\description{
This function attempts to download, compile and install OpenCV
 on the system. This process will take several minutes.
}
\examples{
\dontrun{
installOpenCV()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
