% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{installOpenCV}
\alias{installOpenCV}
\title{Install OpenCV}
\usage{
installOpenCV()
}
\value{
A boolean indicating whether OpenCV was or not installed on the system.
}
\description{
This function will attempt to download, compile and install
 OpenCV on the system. This process will take several minutes.
}
\examples{
\dontrun{
 installOpenCV()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
