% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{installOpenCV}
\alias{installOpenCV}
\title{Install OpenCV}
\usage{
installOpenCV(install_path = defaultOpenCVPath(), batch = FALSE)
}
\arguments{
\item{install_path}{A character string indicating the location at which
OpenCV should be installed. By default, it is the value returned by
\code{\link{defaultOpenCVPath}}.}

\item{batch}{A boolean indicating whether to skip (\code{TRUE}) or not
(\code{FALSE}, the default) the interactive installation dialog. This is
useful when OpenCV needs to be installed in a non-interactive environment
(e.g., during a batch installation on a server).}
}
\value{
A boolean.
}
\description{
This function attempts to download, compile and install OpenCV
 on the system. This process will take several minutes.
}
\examples{
\dontrun{
 installOpenCV()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
