% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{installOpenCV}
\alias{installOpenCV}
\title{Install OpenCV}
\usage{
installOpenCV(batch = FALSE)
}
\arguments{
\item{batch}{A boolean indicating whether to skip (\code{TRUE}) or not (\code{FALSE},
the default) the interactive installation dialog. This is useful when OpenCV needs
to be installed in a non-interactive environment (e.g., during a batch installation
on a server).}
}
\value{
A boolean indicating whether OpenCV was or not installed on the system.
}
\description{
This function will attempt to download, compile and install
 OpenCV on the system. This process will take several minutes.
}
\examples{
\dontrun{
 installOpenCV()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
