% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isCmakeInstalled}
\alias{isCmakeInstalled}
\title{Check Cmake Installation}
\usage{
isCmakeInstalled()
}
\value{
A boolean.
}
\description{
This function checks that Cmake is installed on the system.
}
\examples{
isCmakeInstalled()

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
