% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{isOpenCVInstalled}
\alias{isOpenCVInstalled}
\title{Check OpenCV Installation#' @title Check OpenCV Installation}
\usage{
isOpenCVInstalled()
}
\value{
A boolean indicating whether OpenCV was or not installed on the system.
}
\description{
This functions checks that OpenCV is installed within the R
 library.
}
\examples{
isOpenCVInstalled()

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
