% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{opencvConfig}
\alias{opencvConfig}
\title{C/C++ configuration options}
\usage{
opencvConfig(output = "libs")
}
\arguments{
\item{output}{Either 'libs' for library configuration options or 'cflags' for
C/C++ configuration flags.}
}
\value{
A concatenated character string (with \code{\link{cat}}) of the
 configuration options.
}
\description{
Determines the configuration options for compiling C/C++-based
 packages against OpenCV installed by \code{\link{ROpenCV}}.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

