% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{opencvVersion}
\alias{opencvVersion}
\title{OpenCV Version}
\usage{
opencvVersion()
}
\value{
A character string.
}
\description{
This function determines the version of OpenCV installed within
 R.
}
\examples{
if (isOpenCVInstalled()) {
  opencvVersion()
}

}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
