% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencv.R
\name{opencvVersion}
\alias{opencvVersion}
\title{OpenCV version}
\usage{
opencvVersion()
}
\value{
A character string with the version of OpenCV installed by
 \code{\link{ROpenCV}}.
}
\description{
Determines the version of OpenCV installed within R.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}

