% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROpenCVLite.R
\docType{package}
\name{ROpenCVLite-package}
\alias{ROpenCVLite}
\alias{ROpenCVLite-package}
\title{ROpenCVLite: A package to install OpenCV within R}
\description{
This is a utility package that installs OpenCV within R for use
 by other packages. This Lite version does not contain any of the contributed
 modules of OpenCV.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
